outpout category image on Asset page, refactor Asset components
