+++ client/src/Routes.tsx
@@ -3,7 +3,6 @@ import { Route, Switch } from 'react-router-dom'
 import withTracker from './hoc/withTracker'
 
 import About from './routes/About'
-import Details from './routes/Details/'
 import Home from './routes/Home'
 import NotFound from './routes/NotFound'
 import Publish from './routes/Publish/'
@@ -13,6 +12,7 @@ import History from './routes/History'
 import Channels from './routes/Channels'
 import Styleguide from './routes/Styleguide'
 
+import Asset from './components/templates/Asset'
 import Channel from './components/templates/Channel'
 
 const Routes = () => (
@@ -22,7 +22,7 @@ const Routes = () => (
         <Route component={withTracker(About)} path="/about" />
         <Route component={withTracker(Publish)} path="/publish" />
         <Route component={withTracker(Search)} path="/search" />
-        <Route component={withTracker(Details)} path="/asset/:did" />
+        <Route component={withTracker(Asset)} path="/asset/:did" />
         <Route component={withTracker(Faucet)} path="/faucet" />
         <Route component={withTracker(History)} path="/history" />
         <Route component={withTracker(Channels)} exact path="/channels" />
