+++ client/src/components/templates/Asset/AssetDetails.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables';
+@import '../../../styles/variables';
 
 .metaPrimary {
     margin-bottom: $spacer;
