+++ client/src/components/templates/Asset/AssetFile.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables';
+@import '../../../styles/variables';
 
 .buttonMain {
     margin: auto;
