+++ client/src/components/templates/Asset/AssetFile.tsx
@@ -1,9 +1,9 @@
 import React, { PureComponent } from 'react'
 import { Logger, DDO, File } from '@oceanprotocol/squid'
 import filesize from 'filesize'
-import Button from '../../components/atoms/Button'
-import Spinner from '../../components/atoms/Spinner'
-import { User } from '../../context'
+import Button from '../../atoms/Button'
+import Spinner from '../../atoms/Spinner'
+import { User } from '../../../context'
 import styles from './AssetFile.module.scss'
 import ReactGA from 'react-ga'
