+++ client/src/components/templates/Asset/AssetFilesDetails.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables';
+@import '../../../styles/variables';
 
 .files {
     text-align: center;
