+++ client/src/components/templates/Asset/AssetFilesDetails.test.tsx
@@ -4,8 +4,8 @@ import React from 'react'
 import { render } from 'react-testing-library'
 import { DDO } from '@oceanprotocol/squid'
 import AssetFilesDetails from './AssetFilesDetails'
-import { User } from '../../context'
-import { userMockConnected } from '../../../__mocks__/user-mock'
+import { User } from '../../../context'
+import { userMockConnected } from '../../../../__mocks__/user-mock'
 
 describe('AssetFilesDetails', () => {
     it('renders without crashing', () => {
