+++ client/src/components/templates/Asset/index.test.tsx
