+++ client/src/components/templates/Channel.tsx
@@ -4,7 +4,7 @@ import { History } from 'history'
 import Spinner from '../../components/atoms/Spinner'
 import Route from '../../components/templates/Route'
 import { User } from '../../context'
-import Asset from '../../components/molecules/Asset'
+import AssetTeaser from '../molecules/AssetTeaser'
 import Pagination from '../../components/molecules/Pagination'
 import styles from './Channel.module.scss'
 import Content from '../../components/atoms/Content'
@@ -101,7 +101,7 @@ export default class Channel extends PureComponent<ChannelProps, ChannelState> {
         ) : this.state.results && this.state.results.length ? (
             <div className={styles.results}>
                 {this.state.results.map((asset: any) => (
-                    <Asset key={asset.id} asset={asset} />
+                    <AssetTeaser key={asset.id} asset={asset} />
                 ))}
             </div>
         ) : (
